client: Break validate command out into its own file
